pyproject.toml
[tool.poetry]
name = "knitting-backend"
version = "0.1.0"
description = ""
authors = ["Lemyr27 <sander.96@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
sqlalchemy = {extras = ["async"], version = "^2.0.36"}
uvicorn = "^0.32.0"
pydantic-settings = "^2.6.1"
asyncpg = "^0.30.0"
bcrypt = "^4.2.1"
pyjwt = "^2.10.0"
pydantic = {extras = ["email"], version = "^2.10.1"}


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
greenlet = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
